Fix spreadsheet header
The header is able to show a general dropdown menu when it is sent by
the user. Since the repository was updated to use ViewComponent Slots,
the dropdown was not shown anymore. This beacuse the method method that
condition the dropdown was pointing to a slot of a parent insted of the
slot in the header itself.